Walking the inclusion line
Honestly there are many dozens of huge projects we're learning from, it doesn't make sense to name only a few.

The namecheck I wanted to do happens elsewhere in this doc, where we acknowledge them as direct governance resources. This is because e.g. Kubernetes governance is an evolution of the massively scaled projects of the past.