Clarify what requirement #1 applies to (#154)
After feedback, it's clear that we need to clarify that these requirements (esp requirement #1) apply to the service itself, and not tooling dependencies or service dependencies. 

During the deployment or operating of a service ... many entire stacks of proprietary services or proprietary tooling may be involved (one might use Travis CI or use GitHub CI or proprietary monitoring service, or alerting service, etc. ). It should be possible to meet these requirements without open sourcing an entire constellation of services and tooling that are some how involved.

@danmcp Thanks for this feedback. Can you cross check if the changes resolve the ambiguity?