open-source-services.md: Use word 'dependencies' for clarity
Use the word 'dependencies' instead of 'requirements' for clarity
in the FAQ. What we really mean here is the things that the Service
depends on. The use of the word 'requirements' is already significantly
overloaded here, by referring to the #1 and #2 requirements themselves.

In addition, others have confused these with /product requirements/ or
/testing requirements/. None of these are implied here.

Use of the word 'depndencies' makes this much clearer.